gh-communication/src/main/java/com/infnet/ghcommunication/service/CommunicationServiceImpl.java
package com.infnet.ghcommunication.service;

import org.springframework.stereotype.Service;
import com.infnet.ghcommunication.dto.CommunicationDTO;

@Service
public class CommunicationServiceImpl {

    public CommunicationDTO sendMessage(CommunicationDTO communicationDTO) {
        // Implementar lógica para enviar uma mensagem
        // Converter MessageDto para entidade de domínio, salvar no banco de dados e retornar o MessageDto resultante
        return null;
    }

    public CommunicationDTO getMessageById(Long communicationDTO) {
        // Implementar lógica para buscar uma mensagem pelo ID e retornar o MessageDto correspondente
        return null;
    }

    public void deleteMessage(Long communicationDTO) {
        // Implementar lógica para excluir uma mensagem pelo ID
    }
}
